requirements.txt
anyio==3.6.2
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
docx2txt==0.8
elastic-transport==8.4.0
elasticsearch==8.5.0
fastapi==0.86.0
h11==0.14.0
httptools==0.5.0
idna==3.4
lxml==4.9.1
mysql-connector-python==8.0.31
protobuf==3.20.1
pydantic==1.10.2
PyPDF2==2.11.1
python-docx==0.8.11
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
